[feat] implement get dashboard feature
- seperate Chart Container from DashboardTab to use susepnse
- when year is changed, send api
